+++ web/share-cart/app/routes/_index.tsx
@@ -1,8 +1,9 @@
 import { type MetaFunction, json } from "@remix-run/node";
-import { Form, useLoaderData } from "@remix-run/react";
+import { useLoaderData } from "@remix-run/react";
 import createClient from "openapi-fetch";
 import type { paths } from "~/models/schema";
-import Login from "~/routes/login";
+import Login from "~/routes/auth.login";
+import Logout from "~/routes/auth.logout";
 import { authenticator } from "~/service/auth.server";
 import { config } from "~/service/config.server";
 
@@ -38,9 +39,7 @@ export default function Index() {
 						<p className="text-red-700">You are in development mode</p>
 					)}
 					<Login />
-					<Form action="/auth/logout" method="post">
-						<button type="submit">Logout with Auth0</button>
-					</Form>
+					<Logout />
 					{data.jwt ? (
 						<div>
 							<p className="text-green-700">You are logged in</p>
